# docs/swagger.yaml
basePath: /
definitions:
  api.FavoriteRequest:
    properties:
      product_id:
        type: integer
    type: object
  api.LocationRequest:
    properties:
      location_address:
        type: string
      location_address_ru:
        type: string
      location_name:
        type: string
      location_name_ru:
        type: string
    type: object
...
  models.CartRequest:
    properties:
      count:
        type: integer
      product_id:
        type: integer
      size_id:
        type: integer
      thumbnail_id:
        type: integer
    type: object
...
  models.UpdateProfileRequest:
    properties:
      full_name:
        type: string
      phone:
        type: string
    type: object
...
host: 192.168.55.42:8080
info:
  contact: {}
  description: API for an e-commerce platform with OTP-based authentication and product/market
    management.
  title: E-commerce API
  version: "1.0"
paths:
  /api/cart:
    delete:
      description: Deletes all cart entries for the authenticated user. Requires user
        JWT authentication.
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Clear cart
      tags:
      - Cart
    get:
      description: Retrieves the user's cart grouped by cart order and markets. Requires
        user JWT authentication.
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get user cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Adds or updates a product in the user's cart for a market under
        a single cart order. Requires user JWT authentication.
      parameters:
      - description: Cart entry details
        in: body
        name: cart
        required: true
        schema:
          $ref: '#/definitions/models.CartRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Add to cart
      tags:
      - Cart
...
  /api/favorites:
    get:
      description: Retrieves a paginated list of favorite products. Requires JWT authentication.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get user's favorite products
      tags:
      - Favorites
    post:
      consumes:
      - application/json
      description: Adds or removes a product from favorites. Requires JWT authentication.
      parameters:
      - description: Product ID
        in: body
        name: favorite
        required: true
        schema:
          $ref: '#/definitions/api.FavoriteRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Toggle favorite product
      tags:
      - Favorites
  /api/locations:
    get:
      description: Retrieves all locations for the authenticated user.
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get user locations
      tags:
      - Locations
    post:
      consumes:
      - application/json
      description: Adds a new location with name and address for the authenticated
        user.
      parameters:
      - description: Location details
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/api.LocationRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Add location
      tags:
      - Locations
...
  /api/profile:
    get:
      description: Retrieves the full_name, phone, and id of the authenticated user.
        Requires user JWT authentication.
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Updates the full_name and/or phone of the authenticated user. Requires
        user JWT authentication.
      parameters:
      - description: Profile update details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Profile
...
  /api/user-orders:
    get:
      description: Retrieves all orders for the market admin's market, optionally
        filtered by status. Requires market admin JWT authentication.
      parameters:
      - description: Order status (pending, delivered, cancelled)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get market admin orders
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication (e.g., "Bearer <token>")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"